1. Rearrange unit test files directory structure\n 2. Move sample files to separate folder in spec\n 3. Add tests for LargeFileUploadUtil\n 3.Clean browserfiy action
