+++ .jshintrc
@@ -1,0 +1,3 @@
+{
+  "esnext": "true"
+}
