+++ browser-wrapper.js
@@ -1,4 +1,4 @@
 // make MicrosoftGraph globally accessible
 // MicrosoftGraph.api() can be called directly
 window.MicrosoftGraph = require('./lib/src/index.js');
-Object.assign(window, require("./lib/src/LargeFileUploadTask.js"));
+window.LargeFileUploadTask = require("./lib/src/LargeFileUploadTask.js").LargeFileUploadTask;
