+++ spec/secrets.sample.ts
@@ -1,0 +1,1 @@
+export const AccessToken = "";
