+++ spec/types/miscellaneous.ts
@@ -14,7 +14,7 @@
 */
 
 import { assert } from 'chai'
-import { getClient, randomString } from "./test-helper"
+import { getClient, randomString } from "../test-helper"
 import * as fs from "fs";
